+++ emulator/android/src/main/kotlin/com/bilt/pos/emulator/MainActivity.kt
@@ -1,0 +1,14 @@
+package com.bilt.pos.emulator
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            EmulatorApp(sdkInfo = SdkProbe.describe())
+        }
+    }
+}